requirements.txt
Flask
gunicorn
sqlalchemy
pymysql
python-dotenv
Flask-PyMongo
bcrypt
mysql-connector-python
python-dotenv
firebase-admin